requirements.txt
python-telegram-bot==20.3
websocket-client==1.6.4
pandas==2.1.3
numpy==1.24.3
ta==0.10.2
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
